% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffbot.R
\name{get_analyse}
\alias{get_analyse}
\title{Get data about a URL from Diffbot, using the analyse API}
\usage{
get_analyse(url, api_key, mode = NULL, fallback = NULL, fields = NULL,
  discussion = TRUE, timeout = NULL, callback = NULL)
}
\arguments{
\item{url}{The URL to analyse.}

\item{api_key}{Your API key (also called token in the diffbot documenation).}

\item{mode}{Setting to NULL uses the analyse API, where Diffbot tries to
dermine the page type and analyse accordingly. Other options are: 'article',
 'discussion', 'image', 'product' or 'video'.}

\item{fallback}{If analyse API can't find the correct page type it will use this as a default. Options
are: 'article', 'discussion', 'image', 'product' or 'video'.}

\item{fields}{Depends on the API used. See each APIs documatation page for details.}

\item{discussion}{Set to FALSE to not extract comments for articles and products.}

\item{timeout}{How long to wait before timing out request for URL's content.}

\item{callback}{Used for jsonp requests. See documentation.}
}
\value{
An object of class 'diffbot_api'; a list of information returned from Diffbot.
}
\description{
Get data for a URL from the analyse API.
For more information see \href{https://www.diffbot.com/dev/docs/}{Diffbot's documenation}.
}
\examples{
\dontrun{
api_key <- '12345'
url <- 'http://techcrunch.com/2012/05/31/diffbot-raises-2-million-seed-round-for-web-content-extraction-technology/'

get_diffbot(url, api_key)
get_diffbot(url, api_key, fallback = 'article', discussion = FALSE)
}
}

