% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffbot.R
\name{get_image}
\alias{get_image}
\title{Get data about a URL from Diffbot, using the image API}
\usage{
get_image(url, api_key, fields = NULL, timeout = NULL, callback = NULL)
}
\arguments{
\item{url}{The URL to analyse.}

\item{api_key}{Your API key (also called token in the diffbot documenation).}

\item{fields}{Extra fields, as a vector. Options are: links, meta, querystring, breadcrumb, displayWidth, displayHeight.}

\item{timeout}{How long to wait before timing out request for URL's content.}

\item{callback}{Used for jsonp requests. See documentation.}
}
\value{
An object of class 'diffbot_api'; a list of information returned from Diffbot.
}
\description{
Get data for a URL from the image API.
For more information see \href{https://www.diffbot.com/dev/docs/}{Diffbot's documenation}.
}
\examples{
\dontrun{
api_key <- '12345'
url <- 'http://techcrunch.com/2012/05/31/diffbot-raises-2-million-seed-round-for-web-content-extraction-technology/'

get_diffbot(url, api_key)
get_diffbot(url, api_key, fallback = 'article', discussion = FALSE)
}
}

