% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_diffbot.R
\name{parse_diffbot}
\alias{parse_diffbot}
\title{Parse a 'diffbot_api' object into a dataframe.}
\usage{
parse_diffbot(diffbot_output, include = NULL)
}
\arguments{
\item{diffbot_output}{A object of class diffbot_api.}

\item{include}{NULL for just the page data, 'images' to include
image data or 'tags' to include tags data.}
}
\value{
A dataframe. If just the page data this will be a one row dataframe.
Otherwise will either have a row for each image or each tag.
}
\description{
Take the results of get_diffbot and parse into dataframe.
Optionally include tag data or image data in final frame.
}
\examples{
\dontrun{
url <- 'http://techcrunch.com/2012/05/31/diffbot-raises-2-million-seed-round-for-web-content-extraction-technology/'
api_key <- '12345'

t <- get_diffbot(url, api_key)

parse_diffbot(t)
parse_diffbot(t, 'tags')
}

}

