% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffbot.R
\name{print.diffbot_api}
\alias{print.diffbot_api}
\title{Print a diffbot API object}
\usage{
\method{print}{diffbot_api}(diffbot_return)
}
\arguments{
\item{diffbot_return}{An object of class 'diffbot_api'; a list of information returned from Diffbot.}
}
\value{

}
\description{
Print a diffbot API object
}
\examples{
\dontrun{
api_key <- '12345'
url <- 'http://techcrunch.com/2012/05/31/diffbot-raises-2-million-seed-round-for-web-content-extraction-technology/'

get_diffbot(url, api_key)
}
}

